requirements.txt
streamlit>=1.24.0
ollama>=0.1.5
requests>=2.31.0
python-dotenv>=1.0.0
sqlparse>=0.4.4
pandas>=2.0.0